corrected EOpow

+++ src/main/java/org/eolang/EOfloat.java
@@ -141,7 +141,7 @@ public class EOfloat extends EOObject {
      * @return An object representing the  {@code exponent} free attribute)th power of this float
      */
     public EOfloat EOpow(EOObject exponent){
-        return new EOfloat(Math.pow(this.value, exponent._getData().toInt()));
+        return new EOfloat(Math.pow(this.value, exponent._getData().toFloat()));
     }
 
 }
